Refactor + linking float[]

+++ Java/core/src/main/java/com/amazon/randomcutforest/state/store/LeafStoreState.java
@@ -26,7 +26,7 @@ public class LeafStoreState {
 
     public int[] pointIndex;
     public short[] parentIndex;
-    public int[] mass;
+    public short[] mass;
     public short[] freeIndexes;
 
     public LeafStoreState() {
