+++ Java/core/src/main/java/com/amazon/randomcutforest/tree/CompactNodeView.java
@@ -20,14 +20,14 @@ import java.util.Set;
 
 public class CompactNodeView implements INodeView {
     final AbstractCompactRandomCutTree tree;
-    short currentNodeOffset;
+    int currentNodeOffset;
 
-    public CompactNodeView(AbstractCompactRandomCutTree tree, short initialNodeIndex) {
+    public CompactNodeView(AbstractCompactRandomCutTree tree, int initialNodeIndex) {
         this.tree = tree;
         this.currentNodeOffset = initialNodeIndex;
     }
 
-    public void setCurrentNodeIndex(short newOffset) {
+    public void setCurrentNodeIndex(int newOffset) {
         currentNodeOffset = newOffset;
     }
